requirements.txt
tweepy==4.10.0
alt-profanity-check==1.1.2